Animation and movement update.
'check' directions and 'hit' directions are now calculated based on the next, previous and opposite directions.
Added a game setting for configuring the number of sprite frames for cats and mice.
Updated PositionItem so that it can take an optional x and y modifier to move by a certain number of pixels rather than a big chunk of pixels calculated by grid size and x y coordinates of items.
Updated Animate so that it calls PositionItem when moving the elements, so now the code to position something on the page is only written once.
Fixed Play method so that it doesn't increment the mouse and cat animation frame before calling the Animate method.
Reverting back to using requestAnimationFrame for the animation loop, but it means we need to find a new way to 'Dash' when playing (probably halve the mouseMoveInterval and catMoveInterval would work).
Added reactive template helpers for the playback buttons that react to game.state.